mapping, update/save via reflection + annotations

+++ src/main/java/fr/frogdevelopment/assoplus/dao/CommonDao.java
@@ -13,7 +13,7 @@ public interface CommonDao<E> {
 
 	List<E> getAll();
 
-	E getById(Long identifiant) ;
+	E getById(Integer identifiant) ;
 
 	List<E> getAllOrderedBy(String propertyName);
