+++ src/main/java/fr/frogdevelopment/assoplus/dto/LicenceDto.java
@@ -5,6 +5,7 @@
 package fr.frogdevelopment.assoplus.dto;
 
 import fr.frogdevelopment.assoplus.entities.Licence;
+import javafx.beans.property.SimpleIntegerProperty;
 import javafx.beans.property.SimpleLongProperty;
 import javafx.beans.property.SimpleSetProperty;
 import javafx.beans.property.SimpleStringProperty;
@@ -14,7 +15,7 @@ import org.apache.commons.lang3.builder.ToStringBuilder;
 
 public class LicenceDto implements ReferenceDto {
 
-	private SimpleLongProperty id = new SimpleLongProperty();
+	private SimpleIntegerProperty id = new SimpleIntegerProperty();
 	private SimpleStringProperty code = new SimpleStringProperty();
 	private SimpleStringProperty label = new SimpleStringProperty();
 	private SimpleSetProperty<OptionDto> options = new SimpleSetProperty<>(FXCollections.observableSet());
@@ -39,15 +40,15 @@ public class LicenceDto implements ReferenceDto {
 		return bean;
 	}
 
-	public Long getId() {
+	public Integer getId() {
 		return id.get();
 	}
 
-	public SimpleLongProperty idProperty() {
+	public SimpleIntegerProperty idProperty() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id.set(id);
 	}
